modules/locations-producer/requirements.txt
grpcio-tools==1.41.0
kafka-python==2.0.2
wheel==0.37.0